Fix analyzer errors by separating TagFilters model and updating imports
